requirements.txt
requests
pyodbc
pandas
snowflake-connector-python
python-dotenv
